refactor: Modularize Home component
